Copy the input arguments for async functions locally so that we are dependent on the caller maintaining the lifetime until the async function completes.  Also a bunch of cleanup working and updating to the latest XSAPI nuget package.

+++ CSharpSource/Source/api/TitleStorage/TitleStorageBlobResult.cs
@@ -37,7 +37,7 @@ namespace Microsoft.Xbox.Services.TitleStorage
     internal struct XSAPI_TITLE_STORAGE_BLOB_RESULT
     {
         public IntPtr blobBuffer;
-        public UInt64 cbBlobBuffer;
+        public UInt32 cbBlobBuffer;
         public IntPtr blobMetadata;
     }
 }
